backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
transformers
torch
